fix(Reset): Value not being reset correctly
Prior to this fix when the value of the given ng-model was set to null, if the
user selected the same time as before, the ng-model would not be update.

Calling `element.timepicker('getTime')` after setting the timepicker's value to
null appears to reset it and update the ng-model for all selections correctly.